Movi custom_row_generator() y RowGeneratorTest a distintos modulos nuevos
